Important bug fix: tranpose the contour plot for optimal scaling. In response to Reviewer#1's comment (second round).
plt.contourf() expects a shape of (y, x), where the second dimension is
the used for the x axes. Similar to (lat, lon) for plt.pcolormesh().

This changes the estimate of optimal dx/dz from 1000 to 500, and also more
clearly shows the where the contour lines and solid lines are parallel.

Signed-off-by: JiaweiZhuang <jiaweizhuang@g.harvard.edu>